Refactor dns lists into ACLs
Mostly the same as before except an ACL has a `type`, which can be
DNS or a flat file.

Also reimplement the default user authz setting, which was dropped in
2.0 due to lack of use.

+++ yasl.c
@@ -324,7 +324,7 @@ yaslupdatelen(yastr str) {
  * foo bar "newline are supported\n" and "\xff\x00otherstuff"
  */
 yastr *
-yaslsplitargs(const char * line, int * argc) {
+yaslsplitargs(const char * line, size_t * argc) {
 	if (!line || !argc) { return NULL; }
 
 	const char * p = line;
